refactor(feeds): relocate HttpCache to Feeds namespace
Relocate HttpCache from the single-file snowflake app/web/http/ directory
into Html2rss::Web::Feeds::HttpCache (app/web/feeds/http_cache.rb) to
co-locate it directly within the feeds domain layer.
Remove http from collapsed_web_dirs in boot.rb.
